refactor(spec): split lintSpec layer and scale rule modules
Extract layer-scoped and scale-level lint rules from the lintSpec orchestrator
so each concern can evolve independently. Guard non-record scale entries so
lint never throws on schema-invalid specs. Catalog source-scan covers both
rule modules. Public lintSpec surface is unchanged.
